fix:fix bug where phone field clears on accepting terms
