+++ frontend/src/components/profile/TextMaskCustom.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import MaskedInput from "react-text-mask";
+
+function TextMaskCustom(props) {
+  const { inputRef, ...other } = props
+
+  return (
+    <MaskedInput
+      {...other}
+      ref={(ref) => {
+        inputRef(ref ? ref.inputElement : null)
+      }}
+      mask={['(', '+', /[1-9]/, /\d/, ')', ' ', /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/, /\d/]}
+      placeholderChar={'\u2000'}
+      showMask
+    />
+  )
+}
+
+export default TextMaskCustom;
